refactor: 카카오맵 관련 코드 최상단 페이지로 이동
